src/requirements/base.txt
asgiref==3.8.1
Django==5.0.7
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.6.0

# Black
black==24.4.2
click==8.1.7
colorama==0.4.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2

# CORS
django-cors-headers==4.3.1

# Admin
django-admin-rangefilter==0.13.1

# Celery
# amqp==5.2.0
# billiard==4.2.0
# kombu==5.4.0
# vine==5.1.0
# celery==5.4.0
# click-didyoumean==0.3.1
# click-plugins==1.1.1
# click-repl==0.3.0
# python-dateutil==2.9.0.post0
# prompt_toolkit==3.0.47
# six==1.16.0
# wcwidth==0.2.13
# async-timeout==4.0.3
# redis==5.0.8
